Move unreal log panel classes to tool window package
