--- src/rider/main/kotlin/com/jetbrains/rider/plugins/unreal/toolWindow/ToolWindowUtil.kt
@@ -1,5 +1,0 @@
-package com.jetbrains.rider.plugins.unreal.toolWindow
-
-internal const val TIME_WIDTH = 29
-internal const val VERBOSITY_WIDTH = 12
-internal const val CATEGORY_WIDTH = 20
