+++ src/rider/main/kotlin/com/jetbrains/rider/plugins/unreal/toolWindow/log/UnrealLogConsoleActionsPostProcessor.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.rider.plugins.unreal
+package com.jetbrains.rider.plugins.unreal.toolWindow.log
 
 import com.intellij.execution.ExecutionBundle
 import com.intellij.execution.actions.ClearConsoleAction
