+++ src/rider/main/resources/META-INF/plugin.xml
@@ -171,7 +171,7 @@ Path to old plugin:</code></pre>
   <extensions defaultExtensionNs="com.intellij">
     <statusBarWidgetFactory implementation="com.jetbrains.rider.plugins.unreal.ui.UnrealStatusBarWidget" order="after readOnlyWidget, before fatalErrorWidget" />
     <applicationConfigurable groupId="language" id="UnrealLinkOptionsId" provider="com.jetbrains.rider.settings.UnrealLinkSettingsConfigurableProvider" />
-    <consoleActionsPostProcessor implementation="com.jetbrains.rider.plugins.unreal.UnrealLogConsoleActionsPostProcessor" />
+    <consoleActionsPostProcessor implementation="com.jetbrains.rider.plugins.unreal.toolWindow.log.UnrealLogConsoleActionsPostProcessor" />
     <notificationGroup id="Unreal Editor connection is out of sync" displayType="BALLOON"/>
   </extensions>
